Working tests for MyProfile
