+++ app/src/androidTest/java/com/h/pixeldroid/SwipeTest.kt
@@ -35,6 +35,6 @@ class SwipeTest {
     @Test
     fun swipingRightOnHomepageShowsSettings() {
         onView(withId(R.id.view_pager)).perform(swipeLeft()).perform(swipeLeft()).perform(swipeLeft()).perform(swipeLeft())
-        onView(withId(R.id.nbFollowers)).check(matches(isDisplayed()))
+        onView(withId(R.id.nbFollowersTextView)).check(matches(isDisplayed()))
     }
 }
